// frontend/src/api/generated/models/area-detail.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { AreaType } from './area-type';
// May contain unused imports in some cases
// @ts-ignore
import type { StorageCondition } from './storage-condition';

/**
 * 
 * @export
 * @interface AreaDetail
 */
export interface AreaDetail {
    /**
     * 
     * @type {number}
     * @memberof AreaDetail
     */
    'id': number;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'areaCode': string;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'areaName': string;
    /**
     * 
     * @type {number}
     * @memberof AreaDetail
     */
    'warehouseId': number;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'warehouseCode': string;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'warehouseName'?: string;
    /**
     * 
     * @type {number}
     * @memberof AreaDetail
     */
    'buildingId': number;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'buildingCode': string;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'buildingName'?: string;
    /**
     * 
     * @type {StorageCondition}
     * @memberof AreaDetail
     */
    'storageCondition': StorageCondition;
    /**
     * 
     * @type {AreaType}
     * @memberof AreaDetail
     */
    'areaType': AreaType;
    /**
     * 
     * @type {boolean}
     * @memberof AreaDetail
     */
    'isActive': boolean;
    /**
     * 
     * @type {number}
     * @memberof AreaDetail
     */
    'version': number;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'createdAt': string;
    /**
     * 
     * @type {string}
     * @memberof AreaDetail
     */
    'updatedAt': string;
}